Module aws_sdk_migrationhubconfig::error

Modules

access_denied_exception

See AccessDeniedException

dry_run_operation

See DryRunOperation

internal_server_error

See InternalServerError

invalid_input_exception

See InvalidInputException

service_unavailable_exception

See ServiceUnavailableException

throttling_exception

See ThrottlingException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

CreateHomeRegionControlError

DescribeHomeRegionControlsError

DryRunOperation

Exception raised to indicate that authorization of an action was successful, when the

GetHomeRegionError

InternalServerError

Exception raised when an internal, configuration, or dependency error is

InvalidInputException

Exception raised when the provided input violates a policy constraint or is entered in the

ServiceUnavailableException

Exception raised when a request fails due to temporary unavailability of the

ThrottlingException

The request was denied due to request throttling.

Enums

CreateHomeRegionControlErrorKind

DescribeHomeRegionControlsErrorKind

GetHomeRegionErrorKind